+++ tests/Dummy/RegisterStatesFromCustomDirectories/Directory1/StateA.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Spatie\ModelStates\Tests\Dummy\RegisterStatesFromCustomDirectories\Directory1;
+
+use Spatie\ModelStates\Tests\Dummy\RegisterStatesFromCustomDirectories\RegisterStatesFromCustomDirectories;
+
+class StateA extends RegisterStatesFromCustomDirectories
+{
+}