+++ tests/Dummy/TestModelWithCustomDirectories.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Spatie\ModelStates\Tests\Dummy;
+
+use Illuminate\Database\Eloquent\Model;
+use Spatie\ModelStates\Tests\Dummy\RegisterStatesFromCustomDirectories\RegisterStatesFromCustomDirectories;
+
+class TestModelWithCustomDirectories extends TestModel
+{
+    protected $casts = [
+        'state' => RegisterStatesFromCustomDirectories::class,
+    ];
+}